utils.js, makeMovieCards.js 분리
